템플릿에 이름을 표기하도록 함, 자료구조 추가함.
