+++ item.go
@@ -42,17 +42,6 @@ const (
 	TaskLevel5                   // 5 높은난이도
 )
 
-// Comment 자료구조는 코맨트 글을 작성할 때 사용하는 자료구조이다.
-type Comment struct {
-	Date       string `json:"date"`       // 코맨트 등록시간 RFC3339
-	Author     string `json:"author"`     // 작성자
-	Text       string `json:"text"`       // 내용
-	Media      string `json:"media"`      // media 경로
-	MediaTitle string `json:"mediatitle"` // media 제목
-	Stage      string `json:"stage"`      // 코멘트가 달린 리뷰 Stage
-	Frame      int    `json:"frame"`      // 코맨트 프레임수
-}
-
 // Source 자료구조는 글을 작성할 때 사용하는 자료구조이다.
 type Source struct {
 	Date   string // 등록시간 RFC3339
